fix(SC-01): extract client layout shells, make layout.tsx server components
Move all client-side hooks (useState, usePathname, useLocale,
useClinicFeatures, useTenant, useOnboarding) into dedicated shell
components under src/components/layouts/. The actual layout.tsx files
are now thin React Server Components that import and render the shells.
